Clang update script: fix Android compiler-rt builds after r346167
NDK r16 "helpfully" installs libc++ as libstdc++ to make the NDK's Clang use
that by default. However, we use a more recent Clang, and after r346167 that
starts looking for libc++, which then can't be found due to the renaming.

This hacks around the problem by renaming it back. Well, it makes a copy
actually, to support Clangs both before and after. We should remove all this in
the future when we've moved to a newer NDK.

Bug: 902270
Change-Id: Ib69be5faeb0e0b0c8b5b2588fdb66610e30de687
Reviewed-on: https://chromium-review.googlesource.com/c/1319712
Commit-Queue: Reid Kleckner <rnk@chromium.org>
Reviewed-by: Reid Kleckner <rnk@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#605772}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6d756ac85b9a9a4e8516c4e3f62ed801f854d168