AW: add metric for support library client
No change to logic.

This adds an UMA metric to see if the WebViewClient is a
WebViewClientCompat or not. This will only ever be boolean (at least for
the foreseeable future), so this should be safe to keep as a boolean
histogram for now.

R=changwan@chromium.org, isherman@chromium.org

Bug: 901581
Test: N/A (not feasible to test metrics right now)
Change-Id: I9e6ba219d4f5f26f1fec947ef9974549490e2258
Reviewed-on: https://chromium-review.googlesource.com/c/1318748
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Changwan Ryu <changwan@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#605888}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f05f82998906624e693ac296d6da24cd3dd905b0