[UserTimingL3] Count parameter types in measure method
UserTiming L3 will change the L2 API in certain cases. These cases are described
in http://bit.ly/2SZqAKz.

We want to investigate which and how much various use cases will be impacted by
the introduction of L3 API. So we write this CL to do the investigation.

In particular, we will have two UMA histograms to count the input types of
the performance.measure method in the wild. The input types are to be divided
into the following groups:
1. number
2. sub-groups defined by navigation timing names
3. object
4. unprovided
5. undefined or null
6. others

We used to have a similar histogram before (CL/769932), which we used to collected
2,3,6. Though we had removed it, it defined some of the enums that we reuse.


Change-Id: Idfb1d7fc761189563f55e5f5ba7c493e94e5e83a
Reviewed-on: https://chromium-review.googlesource.com/c/1320012
Commit-Queue: Liquan (Max) Gǔ <maxlg@chromium.org>
Reviewed-by: Nicolás Peña Moreno <npm@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#607437}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 18b8e187f3dded8a818e9b0304592814ce25cabc