Enable Lookalike URL navigation suggestions metrics recording by default
The metrics recording was previously enabled via Finch. It's now ready to be enabled for everyone. This CL enables it by default for all clients.

However, the UI is still not ready to be enabled for everyone. Guard the UI behind a new feature name, so as to avoid confusion from past Finch configs.

In short,
- Old clients which receive the Finch config for LookalikeUrlNavigationSuggestions will record metrics.
- All new clients will record metrics.

Bug: 878153
Change-Id: Ief99afec48d879431ec33e4e74e8722478b45ddd
Reviewed-on: https://chromium-review.googlesource.com/c/1317926
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#607458}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 62ec722a196b7157b559ae2ce490f78d04e23d01