Fix WTF::StringImpl in webkit.natvis
When WTF::StringImpl is displayed in the VisualStudio
debugger, and when the string is expanded, each character
should be displayed, but each character is typed as
StringImpl and that code points cannot be seen.

This patch fixes it for both 8bit and 16bit cases.

r603419 (CL:1288930) removed (wchar_t*) because local test
did not show the differences, but it was because the natvis
in the DLL was prioritized over the local natvis. When both
files were fixed, the symptom was fixed properly.

Change-Id: I68ff0f83456a4e99db803f7140b6e0d6a7630bba
Reviewed-on: https://chromium-review.googlesource.com/c/1337671
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#608454}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2616eba53ad866a8b7053c27eb459250579543f4