Re-land: Add multi results in legacy perf dashboard script.
Currently when the script reads multiple results it overrites the
current result with last value read. It does not compute a mean
or standard deviation.

This change updates the script to keep a list of values. It then
computes a mean and standard deviation from the list.

The first attempt at fixing this returned a list of values. This
was incorrect. Also includes a comment an assert to prevent this.

Bug: 900677
Change-Id: Ided17ea36478128e003e6d9317ca5fb15128415a
Reviewed-on: https://chromium-review.googlesource.com/c/1341211
Commit-Queue: Jamie Madill <jmadill@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#609342}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ce4bc3cb864694fe5ed543ec4047d7c639d4e9b0