Add a flag for rounded corners for PIP.
Applying rounded corners to the PIP window means it currently can not be
displayed using a hardware overlay. Add a flag for this so it's easy to
check how the UI looks with and without and for performance measurement.

  enabled, otherwise no rounded corners are applied.

Bug: 905947
Test: rounded corners is applied to the pip window with this flag
Change-Id: I59697c3fc77c3d253b7b70a1d39e7cc137b201ef
Reviewed-on: https://chromium-review.googlesource.com/c/1341447
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Stefan Kuhne <skuhne@chromium.org>
Commit-Queue: Eliot Courtney <edcourtney@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#609584}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e87a9c3911d928b02483d2756b6b23ec64279b10