[Autofill] New StrikeDatabase class/interface for v2
New StrikeDatabase class for the StrikeDatabase v2. It shares the
same existing ProtoDatabase as legacy_strike_database.h/cc which
is used as persistent storage, as well as an in-memory cache.
Virtual functions are also introduced, which are to be implemented
per project.

Bug: 884817
Change-Id: I48c38b17274ec440afd593dc48223b97c2b2d90e
Reviewed-on: https://chromium-review.googlesource.com/c/1334832
Reviewed-by: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Jared Saul <jsaul@google.com>
Commit-Queue: Anne Lim <annelim@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#609887}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8ae6a515cd4857ea611201cad049cd88c2865822