Clean up PasswordForm.FillOnLoad UKM
There was a discussion about whether Chrome should avoid filling on
load if a form has some new-password fields, or if the condition
should be "has no current-password fields".

The PasswordForm.FillOnLoad metric supported a team decision on that.

The decision was to go with the new condition (for the new parser).

Therefore the metric is now marked as obsolete and no longer
collected.

Bug: 895781
Change-Id: I3a153bf706abe12a5621b0a209cb53b6604ce875
Reviewed-on: https://chromium-review.googlesource.com/c/1341540
Commit-Queue: Vaclav Brozek <vabr@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#609965}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6e66f706936cd2ae9334bb5c80760c7b34ba6e2c