Extracting SubmissionIndicatorEvent enum to separate file.
Now SubmissionIndicatorEvent is part of PasswordForm. But for using it
with NewPasswordFormManager it should be a field of FormData. This
CL extracts it to a separate file and to add FormData::submission_event field.
The following CL will implement processing it in NewPasswordFormManaager.

Bug: 831123
Change-Id: If5e886c29fdfa95396827291436ace66679cd83b
Reviewed-on: https://chromium-review.googlesource.com/c/1341910
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Vaclav Brozek <vabr@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Commit-Queue: Vadym Doroshenko <dvadym@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#610806}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 28a8528fd29c96a37f0e1735af2c2b16cbd8ae89