CacheStorage: Record operation histograms for each operation type separately.
Currently there are a number of histograms that are recorded for each
"operation" that is scheduled.  All operations are aggregated into a single
bucket.  This makes it hard to tell if one operation type is having a
problem.  Also, some operation types are much more frequent than others which
further skews the metrics.

This CL breaks out these operation histograms by the type of the operation.
The aggregated histograms are still preserved for detecting infrastructure
level changes and comparing to historical data.

This CL also migrates the OperationDuration and QueueDuration to the
LONG_TIMES histogram type.  We have top level metrics that greatly exceed
the default 10 seconds and we expect some of the specific operations will
need to measure very long values on some platforms.

TBR=jkarlin
    cannot be easily further reduced.  Bug 909894 has been filed as a
    follow-on to reduce the code bloat further.

Binary-Size: Increase is due to a large number of new UMA histograms which
Bug: 908851
Change-Id: Iee607ae8f9f15384058af4977fffcffe027e0dbc
Reviewed-on: https://chromium-review.googlesource.com/c/1352479
Commit-Queue: Ben Kelly <wanderview@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#611912}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 218cb607d1cb13d69f0723a7d2a19290296c4598