Navigation Predictor: Add logic to preconnect to origins with highest preconnect score
At onLoad, Preconnect to the origin that has the highest preconnect
score as long as the score is above the specified threshold.
Currently, due to privacy constraints, this is restricted to
only the document origin.

Effectively, this means that we preconnect to the document origin
only if its preconnect score is more than the specified threshold.
If not, no preconnect action is taken.

The preconnect score of an origin is computed from the score
of all the anchor elements on the webpage that point to that
origin.

Change-Id: Ie3178f352cb5106124e13cb6f7e39b9ecf81174c
Bug: 903945
Reviewed-on: https://chromium-review.googlesource.com/c/1349874
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Commit-Queue: Tarun Bansal <tbansal@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#611955}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f2854a0d56921d4c43a7afe6830c1503b9bf4a80