[Ephemeral Panel] Add histograms for CTR, etc.
Adds histograms to track CTR and closing of the Ephemeral Tab.
When the user makes the panel peek they may or may not actually expand it.
We now track whether it was ever expanded in a boolean histogram, and
the reason it was closed in an enumerated histogram.

BUG=894619, 902122

WIP Ephemeral Tab log when activated.

Change-Id: I27e6a5b36fc0efc0c84298235b51b2df915196f6
Reviewed-on: https://chromium-review.googlesource.com/c/1356033
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Commit-Queue: Donn Denman <donnd@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#612730}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a572518db668f5a2331e0d5bed912496018b3eb4