[Telemetry][CI] Replace --smoke-test-mode with --pageset-repeat=1
--smoke-test-mode's only effect is to set pageset-repeat to 1.
This is only true for benchmark runs. For merge scripts,
that flag has different meaning and does not need to be removed.

Note that Note that we currently support duplicate conflicting flags:
for example
‘run_benchmark ... --pageset-repeat=2 --pageset-repeat=3’
We use the last flag, 3 in the example.

--smoke-test-mode is getting deprecated in
https://chromium-review.googlesource.com/c/catapult/+/1356903


Bug: 894254
Change-Id: I1b0e423167018f6010f66255838ad8a33d6acd76
Reviewed-on: https://chromium-review.googlesource.com/c/1357722
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Reviewed-by: Emily Hanley <eyaich@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Commit-Queue: Caleb Rouleau <crouleau@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#613342}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5067b69c8d75e8667d089aa2cd93bc7bbf4b3999