child user: Add ParentAccessCodeConfig user policy
This policy specifies configuration that is used to generate/verify
Parent Access Code on child user's device.
This policy applies only to child user.
The value of this policy should not be displayed in the UI and
it is marked as 'sensitiveValue' therefore.

Bug: 911326
Test: policy_test_cases.json
Change-Id: Iee3dafce712b67f038cea52d5fc8810a6532e809
Reviewed-on: https://chromium-review.googlesource.com/c/1359893
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Aga Wronska <agawronska@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#614539}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 24fde26e7643d0bb161d27cc1a1b79432e5752e5